requirements.txt
psycopg2-binary
matplotlib
chardet
python-dotenv
